qa/pyproject.toml
[project]
name = "beehive-blog-v3-qa"
version = "0.1.0"
description = "Repository-managed HTTP regression and performance test project for Beehive Blog v3."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28.1,<0.29.0",
  "locust>=2.41.6,<3.0.0",
  "pydantic>=2.11.7,<3.0.0",
  "pytest>=8.4.2,<9.0.0",
  "pytest-xdist>=3.8.0,<4.0.0",
  "python-dotenv>=1.1.1,<2.0.0",
  "pyyaml>=6.0.3,<7.0.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra"

